fix(frontend): move UebersichtTabSkeleton out of the app/ route tree
The dependency-cruiser rule "components-not-to-app-or-api-routes" forbids
src/components/ from importing src/app/ route files. uebersicht-tab.tsx
imported UebersichtTabSkeleton from the staff/[id] route's page-skeleton.tsx,
violating that direction. Move it to components/staff/ instead, imported
by both the component and the route skeleton (the allowed direction).
